fix(char): make character deletion work, and stop inventing a main character
Deleting a character did nothing and said nothing. Two faults, and they
hid each other.

Every id in this exchange is u64, and we wrote or read u32:

  CSDeleteCharacter          u64 type                      (RVA 0xC688D0)
  CSCancelCharacterDelete    the same serialize function
  SCDeleteCharacterResponse  u64 type, u8 status,
                             u64 requestedTime, u64 delay  (RVA 0xC59C20)
  SCCharacterDeleted         u64 type, wstring name        (RVA 0xC59CA0)
  SCCancelCharacterDeleteResponse  u64 type, u8 status     (RVA 0xC59D40)

Reading four bytes where the client writes eight meant the id was never
the one the player clicked, so CharacterManager found no such character
and answered with a rejection - which the client could not read either,
because the response had the same fault. Nothing happened and no error
appeared. On SCDeleteCharacterResponse the short id also shifted the
status and both timestamps by four bytes.

The second fault: character select named characters[0] as the account's
main ("represent") character with success=true, unconditionally. The
client stores that id as THE represent character (RVA 0x4DD270) and its
delete dialog then refuses it with "Must deselect as Main Character
before deleting." - a main character nobody had chosen, reasserted at
every login, and always the same slot.

So the account now owns that choice. characters.represent stores it, at
most one per account; CSRepresentCharacter sets or clears it and
SCRepreSentCharacter answers. Character select sends whoever the player
actually nominated, with success=false when that is nobody, which leaves
the client's id at its default of zero.

SetDeleteCharacter refuses to delete the main character. The client
already refuses it, but that check lives entirely in its UI.

+++ AAEmu.Game/Core/Packets/G2C/SCDeleteCharacterResponsePacket.cs
@@ -3,6 +3,18 @@ using AAEmu.Game.Core.Network.Game;
 
 namespace AAEmu.Game.Core.Packets.G2C;
 
+/// <summary>
+/// Client layout (read at RVA 0xC59C20), which names every field:
+///
+///   type                u64   the character id - EIGHT bytes, not four
+///   deleteStatus        u8
+///   deleteRequestedTime u64   unix seconds
+///   deleteDelay         u64   unix seconds
+///
+/// The id was written as u32, which shifted the status and both timestamps by four bytes. Neither the
+/// confirmation nor the rejection could be read, so a delete request produced no visible reaction at
+/// all - not even an error.
+/// </summary>
 public class SCDeleteCharacterResponsePacket(
     uint characterId,
     byte status,
@@ -15,10 +27,10 @@ public class SCDeleteCharacterResponsePacket(
 
     public override PacketStream Write(PacketStream stream)
     {
-        stream.Write(characterId);
-        stream.Write(status);
-        stream.Write(_deleteRequestedTime);
-        stream.Write(_deleteDelay);
+        stream.Write((ulong)characterId);   // u64 type
+        stream.Write(status);               // u8  deleteStatus
+        stream.Write(_deleteRequestedTime); // u64 deleteRequestedTime
+        stream.Write(_deleteDelay);         // u64 deleteDelay
         return stream;
     }
 }
